coord: extract id allocator into ore
So that it can be reused.
