+++ src/ore/src/id_gen.rs
@@ -15,7 +15,9 @@
 
 //! ID generation utilities.
 
+use std::collections::VecDeque;
 use std::marker::PhantomData;
+use std::sync::Mutex;
 
 /// Manages the allocation of unique IDs.
 #[derive(Debug, Default)]
@@ -35,3 +37,84 @@ impl<Id: From<u64> + Default> Gen<Id> {
 
 /// A generator of u64-bit IDs.
 pub type IdGen = Gen<u64>;
+
+/// Manages allocation of u32 IDs.
+///
+/// Note that the current implementation wastes memory. It would be far more
+/// efficient to use a compressed bitmap, like <https://roaringbitmap.org> or
+/// the hibitset crate, but neither presently supports a fast "find first zero"
+/// operation.
+#[derive(Debug)]
+pub struct IdAllocator(Mutex<IdAllocatorInner>);
+
+#[derive(Debug)]
+struct IdAllocatorInner {
+    next: u32,
+    max: u32,
+    free: VecDeque<u32>,
+}
+
+impl IdAllocator {
+    /// Creates a new `IdAllocator` that will assign IDs between `min` and
+    /// `max`, both inclusive.
+    pub fn new(min: u32, max: u32) -> IdAllocator {
+        IdAllocator(Mutex::new(IdAllocatorInner {
+            next: min,
+            max,
+            free: VecDeque::new(),
+        }))
+    }
+
+    /// Allocates a new ID.
+    ///
+    /// Returns `None` if the allocator is exhausted.
+    pub fn alloc(&self) -> Option<u32> {
+        let mut inner = self.0.lock().expect("lock poisoned");
+        if let Some(id) = inner.free.pop_front() {
+            Some(id)
+        } else {
+            let id = inner.next;
+            if id > inner.max {
+                None
+            } else {
+                inner.next += 1;
+                Some(id)
+            }
+        }
+    }
+
+    /// Releases a new ID back to the pool.
+    ///
+    /// It is undefined behavior to free an ID twice, or to free an ID that was
+    /// not allocated by this allocator.
+    pub fn free(&self, id: u32) {
+        let mut inner = self.0.lock().expect("lock poisoned");
+        inner.free.push_back(id);
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_id_alloc() {
+        let ida = IdAllocator::new(3, 5);
+        assert_eq!(ida.alloc().unwrap(), 3);
+        assert_eq!(ida.alloc().unwrap(), 4);
+        assert_eq!(ida.alloc().unwrap(), 5);
+        ida.free(4);
+        assert_eq!(ida.alloc().unwrap(), 4);
+        ida.free(5);
+        ida.free(3);
+        assert_eq!(ida.alloc().unwrap(), 5);
+        assert_eq!(ida.alloc().unwrap(), 3);
+        match ida.alloc() {
+            Some(id) => panic!(
+                "id allocator returned {}, not expected id exhaustion error",
+                id
+            ),
+            None => (),
+        }
+    }
+}
